+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-- Schema: add `isPublished` and `isMassAssignable` attribute to fields
-- Sequelize: honour `isPublished` in `model.toJSON()` and `isMassAssignable` in `model.set()`
+- Schema: add `isMassAssignable` attribute to fields
+- Sequelize: honour `existsInClient` and `existsInServer` in `model.toJSON()` and `isMassAssignable` in `model.set()`
 
 ## 0.3.0 (2018-4-27)